requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0

# LLM
anthropic>=0.34.0
langchain-anthropic>=0.1.0
langchain-core>=0.2.0

# Document ingestion
pypdf>=4.3.0
python-docx>=1.1.0
unstructured>=0.14.0

# Embeddings & vector store
chromadb>=0.5.0
langchain-community>=0.2.0

# Reporting
reportlab>=4.2.0
plotly>=5.22.0
pandas>=2.2.0
jinja2>=3.1.0

# Dashboard
streamlit>=1.36.0
streamlit-extras>=0.4.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
